Check exist ~/.gitignore_global

+++ gitignore/gitignore.sh
@@ -3,7 +3,7 @@
 $MAC_SETUP_DIR/install.sh gitignore \
   --script-version "" \
   --script-config "
-    rm ~/.gitignore_global
+    [ -f ~/.gitignore_global ] && rm ~/.gitignore_global
     ln -s $MAC_SETUP_DIR/gitignore/gitignorerc ~/.gitignore_global
     ls -la ~/.gitignore_global" \
   $@
